chore(test): Added testing to the new get network route, refactored pytest fixtures to avoid code repetition by shifting configured manager setups to the conftest.py.
