+++ test/networking/test_p2p_networking.py
@@ -11,158 +11,7 @@ logger = cp._logger.getLogger("chimerapy")
 cp.debug()
 
 
-@pytest.fixture
-def graph(gen_node, con_node):
-
-    # Define graph
-    _graph = cp.Graph()
-    _graph.add_nodes_from([gen_node, con_node])
-    _graph.add_edge(src=gen_node, dst=con_node)
-
-    return _graph
-
-
-@pytest.fixture
-def single_node_no_connections_manager(manager, worker, gen_node):
-
-    # Define graph
-    simple_graph = cp.Graph()
-    simple_graph.add_nodes_from([gen_node])
-
-    # Connect to the manager
-    worker.connect(host=manager.host, port=manager.port)
-
-    # Then register graph to Manager
-    assert manager.commit_graph(
-        simple_graph,
-        {
-            "local": ["Gen1"],
-        },
-    )
-
-    return manager
-
-
-@pytest.fixture
-def multiple_nodes_one_worker_manager(manager, worker, graph):
-
-    # Connect to the manager
-    worker.connect(host=manager.host, port=manager.port)
-
-    # Then register graph to Manager
-    assert manager.commit_graph(
-        graph,
-        {
-            "local": ["Gen1", "Con1"],
-        },
-    )
-
-    return manager
-
-
-@pytest.fixture
-def multiple_nodes_multiple_workers_manager(manager, graph):
-
-    worker1 = cp.Worker(name="local", port=0)
-    worker2 = cp.Worker(name="local2", port=0)
-
-    worker1.connect(host=manager.host, port=manager.port)
-    worker2.connect(host=manager.host, port=manager.port)
-
-    # Then register graph to Manager
-    assert manager.commit_graph(graph, {"local": ["Gen1"], "local2": ["Con1"]})
-
-    yield manager
-
-    worker1.shutdown()
-    worker2.shutdown()
-
-
-@pytest.fixture
-def slow_single_node_single_worker_manager(manager, worker, slow_node):
-
-    # Define graph
-    simple_graph = cp.Graph()
-    simple_graph.add_nodes_from([slow_node])
-
-    # Connect to the manager
-    worker.connect(host=manager.host, port=manager.port)
-
-    # Then register graph to Manager
-    assert manager.commit_graph(
-        simple_graph,
-        {
-            "local": ["Slo1"],
-        },
-    )
-
-    return manager
-
-
-@pytest.fixture
-def dockered_single_node_no_connections_manager(dockered_worker, manager, gen_node):
-
-    # Define graph
-    simple_graph = cp.Graph()
-    simple_graph.add_nodes_from([gen_node])
-
-    # Connect to the manager
-    dockered_worker.connect(host=manager.host, port=manager.port)
-
-    # Then register graph to Manager
-    assert manager.commit_graph(
-        simple_graph,
-        {
-            dockered_worker.name: ["Gen1"],
-        },
-    )
-
-    return manager
-
-
-@pytest.fixture
-def dockered_multiple_nodes_one_worker_manager(
-    dockered_worker, manager, gen_node, con_node
-):
-
-    # Define graph
-    simple_graph = cp.Graph()
-    simple_graph.add_nodes_from([gen_node, con_node])
-    simple_graph.add_edge(gen_node, con_node)
-
-    # Connect to the manager
-    dockered_worker.connect(host=manager.host, port=manager.port)
-
-    # Then register graph to Manager
-    assert manager.commit_graph(
-        simple_graph,
-        {
-            dockered_worker.name: ["Gen1", "Con1"],
-        },
-    )
-
-    return manager
-
-
-@pytest.fixture
-def dockered_multiple_nodes_multiple_workers_manager(docker_client, manager, graph):
-
-    worker1 = DockeredWorker(docker_client, name="local")
-    worker2 = DockeredWorker(docker_client, name="local2")
-
-    worker1.connect(host=manager.host, port=manager.port)
-    worker2.connect(host=manager.host, port=manager.port)
-
-    # Then register graph to Manager
-    assert manager.commit_graph(graph, {"local": ["Gen1"], "local2": ["Con1"]})
-
-    yield manager
-
-    worker1.shutdown()
-    worker2.shutdown()
-
-
-#@pytest.mark.repeat(10)
+# @pytest.mark.repeat(10)
 @pytest.mark.parametrize(
     "config_manager, expected_worker_to_nodes",
     [
